Convert "Groups" pages to typescript
